Рефактор собственных исключений и их обработчика

+++ src/main/java/com/delmark/portfoilo/exceptions/ErrorResponse.java
@@ -1,12 +1,13 @@
 package com.delmark.portfoilo.exceptions;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
 public class ErrorResponse {
 
-    private final Long codeID;
+    private final Integer codeID;
 
     private final String message;
 }
